Moved options-parser-util.ts to appdistirbution directory.
